+++ src/lib.rs
@@ -15,7 +15,6 @@ pub mod query;
 pub mod registry;
 pub mod semantic_component;
 pub mod snapshot;
-pub mod weight;
 
 pub mod plugin;
 
@@ -27,13 +26,12 @@ pub mod prelude {
         CommandsExt, SemanticEdit, SemanticPlaybackQueue, Semantics, SemanticsPlugin,
     };
     pub use crate::query::{
-        CompiledEdgeQuery, CompiledTraversalQuery, EdgeQuery, EdgeQueryBuilder, SemanticCommand,
-        SemanticEdge, TraversalQuery, TraversalQueryBuilder,
+        CompiledEdgeQuery, CompiledTraversalQuery, EdgeQuery, EdgeQueryBuilder, EdgeRegistration,
+        SemanticCommand, SemanticEdge, TraversalQuery, TraversalQueryBuilder,
     };
     pub use crate::registry::{Core, SemanticRegistry, SemanticRegistryBatch};
     pub use crate::semantic_component::{SemanticAppExt, SemanticComponents, SemanticWorldExt};
     pub use crate::snapshot::SemanticSnapshot;
-    pub use crate::weight::Weight;
     pub use crate::{kind, semantic_component, semantic_kinds, SemanticComponent};
 }
 
@@ -41,15 +39,14 @@ pub use direction::EdgeDirection;
 pub use error::SemanticError;
 pub use kind::{Kind, KindRegistration};
 pub use query::{
-    CompiledEdgeQuery, CompiledTraversalQuery, EdgeQuery, EdgeQueryBuilder, SemanticCommand,
-    SemanticEdge, TraversalQuery, TraversalQueryBuilder,
+    CompiledEdgeQuery, CompiledTraversalQuery, EdgeQuery, EdgeQueryBuilder, EdgeRegistration,
+    SemanticCommand, SemanticEdge, TraversalQuery, TraversalQueryBuilder,
 };
 pub use registry::{Core, SemanticRegistry, SemanticRegistryBatch};
 pub use semantic_component::{
     SemanticAppExt, SemanticComponent, SemanticComponents, SemanticWorldExt,
 };
 pub use snapshot::SemanticSnapshot;
-pub use weight::Weight;
 
 pub use plugin::{CommandsExt, SemanticEdit, SemanticPlaybackQueue, Semantics, SemanticsPlugin};